+++ bindings/java/build.gradle
@@ -1,38 +1,31 @@
+// Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
     repositories {
         google()
-        
-        // Must specify for android building, as the rust cargo plugin is out of date
-        maven {
-            url "https://plugins.gradle.org/m2/"
-            content {
-                excludeGroup "com.android.tools.build"
-            }
-        }
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.2'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.3.31"
-        classpath 'gradle.plugin.org.mozilla.rust-android-gradle:plugin:0.8.3'
+        classpath "com.android.tools.build:gradle:4.1.0"
+
+        // NOTE: Do not place your application dependencies here; they belong
+        // in the individual module build.gradle files
     }
 }
 
 allprojects {
     repositories {
         google()
-        // Must specify for android building, as the rust cargo plugin is out of date
-        maven {
-            url "https://plugins.gradle.org/m2/"
-            content {
-                excludeGroup "com.android.tools.build"
-            }
-        }
         jcenter()
     }
 }
-// Android deleting, TODO move to examples app
+
 task clean(type: Delete) {
-    delete "${rootDir}/native/bin/"
-    delete "${rootDir}/lib/"
+    delete rootProject.buildDir
+    delete "${rootDir}/examples/android-app/target"
+    delete "${rootDir}/examples/java-app/bin"
+    delete "${rootDir}/examples/java-app/build"
+    delete "${rootDir}/native/build"
+    delete "${rootDir}/native/target"
+    delete "${rootDir}/build"
+    delete "${rootDir}/target"
 }
