+++ bindings/java/examples/android-app/CMakeLists.txt
@@ -1,0 +1,5 @@
+# Sets the minimum version of CMake required to build your native library.
+# This ensures that a certain set of CMake features is available to
+# your build.
+
+cmake_minimum_required(VERSION 3.4.1)
