+++ bindings/java/examples/android-app/src/main/java/org/iota/example/MainActivity.java
@@ -1,0 +1,16 @@
+package org.iota.example;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import android.os.Bundle;
+
+public class MainActivity extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_main);
+
+        new MyApplication(this);
+    }
+}
