+++ bindings/java/examples/java-app/.gitignore
