+++ bindings/java/native/pom.xml
@@ -4,7 +4,32 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>org.iota.client</groupId>
+  <parent>
+      <groupId>org.iota.client</groupId>
+      <artifactId>native-parent</artifactId>
+      <version>0.0.1-SNAPSHOT</version>
+  </parent>
+
+  <name>client.rs native bindings</name>
   <artifactId>native</artifactId>
-  <version>1.0</version>
+  <packaging>jar</packaging>
+
+  <!-- Move output to the same as gradle output so instructions are consistent -->
+  <build>
+    <finalName>${project.artifactId}</finalName>
+    <directory>build</directory>
+
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-jar-plugin</artifactId>
+        <version>2.3.1</version>
+        <configuration>
+          <!-- Move final jar to the libs folder -->
+          <outputDirectory>build/libs</outputDirectory>
+        </configuration>
+      </plugin>
+    </plugins>
+  </build>
 </project>
+
