+++ bindings/java/native/src/java_glue.rs
@@ -1,7 +1,6 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-// Stop yelling at generated code
 #![allow(warnings, clippy::all)]
 
 include!(concat!(env!("OUT_DIR"), "/java_glue.rs"));
