+++ bindings/java/native/src/jni_c_header.rs
@@ -1,7 +1,17 @@
 // Copyright 2020 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-// Stop yelling at generated code
-#![allow(warnings, clippy::all)]
+// deref_nullptr: supressing UB warnings in bindgen generated code. See:
+// https://github.com/rust-lang/rust-bindgen/issues/1651
+#![allow(
+    non_upper_case_globals,
+    dead_code,
+    non_camel_case_types,
+    improper_ctypes,
+    non_snake_case,
+    deref_nullptr,
+    clippy::unreadable_literal,
+    clippy::const_static_lifetime
+)]
 
-include!(concat!(env!("OUT_DIR"), "/jni_c_header.rs"));
+include!(concat!(env!("OUT_DIR"), "/jni_c_headers.rs"));
